CLN: Refactor code
