+++ mitoviz/constants.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: UTF-8 -*-
 # Created by Roberto Preste
+
 NAMES = ["DLOOP", "TF", "RNR1", "TV", "RNR2", "TL1",
          "ND1", "TI", "TQ", "TM", "ND2", "TW", "TA", "TN",
          "TC", "TY", "CO1", "TS1", "TD", "CO2", "TK", "ATP8",
@@ -44,31 +45,3 @@ TEXT_Y = [25.2, 25.2, 25.2, 25.2, 25.2, 25.2, 25.2, 19.2, 25.2, 19.2, 25.2,
           25.2, 19.2, 25.2, 19.2, 25.2, 25.2, 25.2, 19.2, 25.2, 19.2, 25.2,
           25.2, 25.2, 25.2, 25.2, 25.2, 25.2, 25.2, 25.2, 19.2, 25.2, 25.2,
           25.2, 25.2, 25.2, 25.2, 19.2]
-
-thetas = [0.0, 0.2264460739936025, 0.42100368157402374, 0.6151816645542881,
-          0.9241960890820207, 1.234349387410224, 1.430045868791116,
-          1.6246034763715373, 1.6513670106826002, 1.6779407326935845,
-          1.8886323857806748, 2.099324038867765, 2.1253283239785143,
-          2.1522816705896557, 2.1786655803005615, 2.203720803910918,
-          2.5089389824370816, 2.8152960347637155, 2.8418697567746998,
-          2.984608606433702, 3.1277270806928605, 3.180305087814594,
-          3.348858410284266, 3.6269334298992097, 3.788653509566057,
-          3.8672358017985395, 3.9452486571307865, 4.013960709759189,
-          4.331896312390609, 4.60655471060414, 4.630850685014183,
-          4.655526284024383, 5.012942845072122, 5.456534190355484,
-          5.569282696602088, 5.798955579697025, 6.028059025891725,
-          6.053683686402318]
-
-widths = [0.4259388013760637, 0.026953346611141286, 0.3621618685497012,
-          0.026194097410827448, 0.5918347516446376, 0.028471845011768967,
-          0.36292111775001507, 0.026194097410827448, 0.027332971211298206,
-          0.025814472810670532, 0.3955688333635102, 0.025814472810670532,
-          0.026194097410827448, 0.02771259581145513, 0.02505522361035669,
-          0.02505522361035669, 0.58538113344197, 0.027332971211298206,
-          0.025814472810670532, 0.259663226507333, 0.02657372201098437,
-          0.07858229223248235, 0.2585243527068622, 0.2976256865230249,
-          0.025814472810670532, 0.13135011165429417, 0.02467559901019977,
-          0.11274850624660511, 0.5231226990162352, 0.026194097410827448,
-          0.022397851409258256, 0.026953346611141286, 0.6878797754843382,
-          0.19930291508238276, 0.026194097410827448, 0.4331516687790452,
-          0.02505522361035669, 0.026194097410827448]
